Another (final?) refactor of PPU reg handling
I've changed course a few times here, but I think this is the final
iteration (for now) of config register management. Before I was reading
config registers to a `HashMap`, but this made handling register fields
(ie. LCDC fields) clunky to access. In this implementation, I've
completely changed the approach - register values are now stored as
decoded fields in `PPUConfig`, and are encoded/flushed to memory each
cycle.